+++ backend/canisters/user/impl/src/model/mod.rs
@@ -1,6 +1,5 @@
 pub mod communities;
 pub mod community;
-pub mod game_chit_keys;
 pub mod group_chat;
 pub mod group_chats;
 pub mod local_user_index_event_batch;